requirements.txt
numpy>=2.0
scipy>=1.10
matplotlib>=3.7
tqdm>=4.66
arviz>=0.16
pystan>=3.7
httpstan>=4.10
POT>=0.9.3
xarray>=2023.7.0
